requirements.txt
python-barcode==0.15.1
pillow==9.5.0
numpy==1.26.3
opencv-python==4.8.1.78
torch==2.2.1
diffusers==0.25.0
tqdm==4.66.1
